Rename InputModal to FormModal
